Extracted XML(TRX) document creation to Helper class
