+++ 4b/Interfaces/src/TRX/TestDetails.cs
@@ -3,6 +3,7 @@ namespace Interfaces.src.TRX
     public class TestDetails
     {
         public UnitTestResult UnitTestResult { get; set; }
+        
         public UnitTestDefinition UnitTestDefinition { get; set; }
         
         public TestEntry TestEntry { get; set; }
